+++ src/contracts/ValidatorRegistryContract.ts
@@ -1,15 +1,11 @@
 import { inject, injectable } from 'inversify';
-import fs from 'fs';
-import path from 'path';
-import { Contract, ContractInterface, utils } from 'ethers';
+import { Contract, utils } from 'ethers';
 import Settings from '../types/Settings';
 import Blockchain from '../lib/Blockchain';
-import {ContractRegistry} from '@umb-network/toolbox';
+import {ContractRegistry, ABI} from '@umb-network/toolbox';
 
 @injectable()
 class ValidatorRegistryContract {
-  static ABI: ContractInterface = fs.readFileSync(path.resolve(__dirname, './ValidatorRegistryContract.abi.json'), 'utf-8');
-
   contract!: Contract;
 
   constructor(
@@ -21,7 +17,7 @@ class ValidatorRegistryContract {
       .then((validatorRegistryAddress: string) => {
         this.contract = new Contract(
           validatorRegistryAddress,
-          ValidatorRegistryContract.ABI,
+          ABI.validatorRegistryAbi,
           blockchain.provider
         ).connect(blockchain.wallet);
       });